feat: store inventory items by id rather than nested structure
Changes the way that the inventory class stores items such that
everything is in a flat dictionary that is addressed using a UUID. This
simplifies item access makes it easier to prove the independence of two
items.

This change also implement save/load for the currently equipped item
because it just addresses the UUID of said weapon.